- Implement changes in regards to how available years and months are
  fetched
- Added some query optimizations

+++ MixxFit.API/Domain/Entities/WeightEntries/WeightEntry.cs
@@ -9,7 +9,7 @@ namespace MixxFit.API.Domain.Entities.WeightEntries
         public int? FitnessProfileId { get; set; }
         public FitnessProfile? FitnessProfile { get; set; }
 
-        public double Weight { get; set; }
+        public decimal Weight { get; set; }
         public TimeSpan Time { get; set; }
 
         public string? Notes { get; set; }
